backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI companion API with HiMem long-term memory"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "transformers>=4.30.0",
    "tokenizers>=0.13.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "loguru>=0.7.0",
    "pendulum>=3.0.0",
    "parsedatetime>=2.6",
    "rank-bm25>=0.2.2",
    "opensearch-py==3.1.0",
    "faiss-cpu>=1.7.0",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.0",
    "pytz>=2026.1.post1",
    "sqlalchemy>=2.0.48",
    "pyjwt>=2.12.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.27.0"]
